+++ R/utilities.R
@@ -119,16 +119,18 @@ valid_query = function(query) {
 #' @examples
 #'
 #' # Add suffix
-#' append_r_suffix("testing", rlang = TRUE, suffix = "toad")
+#' append_search_term_suffix("testing", rlang = TRUE, suffix = "toad")
 #'
 #' # Retain original search query
-#' append_r_suffix("testing", rlang = FALSE, suffix = "toad")
+#' append_search_term_suffix("testing", rlang = FALSE, suffix = "toad")
 #' @noRd
-append_r_suffix = function(query, rlang = TRUE, suffix = "r programming") {
-  if (rlang)
+append_search_term_suffix = function(query, rlang = TRUE, suffix = "r programming") {
+
+  if (rlang && !is.null(suffix))
     paste(query, suffix)
   else
     query
+
 }
 
 #' Check if in RStudio